Provide two implementations of Release_buffer.
The Lwt_bytes (bigarray-based) implementation seems to require
large amounts of memory for long-running processes. It's hard to
tell if this is a leak or not, but string buffers seem to be less
memory hungry. It should be less performant, however, since more
copies are necessary.

This commit breaks backward compatibility and will only be
included in a new major release.